CommentClassifier Implemented
Here is a preliminar version of the CommentClassifier for multiple classes

+++ Code/classes/CommentDataset.py
@@ -20,7 +20,7 @@ class CommentDataset (Dataset):
         item = self.data.iloc[index]
         comment = str(item.Review)
         comment = comment[:max_len_btweet] if len(comment) > max_len_btweet else comment
-        label = torch.FloatTensor(self.data.iloc[index, 1:])
+        label = torch.LongTensor(self.data.iloc[index, 1:])
         encoding = self.tokenizer.encode_plus(
                                 comment,
                                 add_special_tokens=True,
